table/tableBatch moved to separate files
